Complete RipaAlert, RipaAutocomplete, RipaStopDate tests
