+++ .gitignore
@@ -391,3 +391,5 @@ Functions/__azurite_db_blob_extent__.json
 Functions/__azurite_db_queue__.json
 Functions/__azurite_db_queue_extent__.json
 Functions/__azurite_db_table__.json
+
+__snapshots__/
